requirements.txt
aioamqp==0.14.0
aioredis==1.3.1
asgiref==3.3.0
asn1crypto==1.4.0
async-timeout==3.0.1
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
certifi==2020.6.20
cffi==1.14.3
channels-rabbitmq==1.2.1
chardet==3.0.4
constantly==15.1.0
cryptography==3.1.1
daphne==2.5.0
Django==3.1.2
django-channels==0.7.0
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
msgpack==1.0.0
oauthlib==3.1.0
pamqp==2.3.0
pkg-resources==0.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.11
zope.interface==5.1.2
